model/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.1
requests>=2.32.2
opencv-python==4.8.1.78
torch==2.2.0
torchvision==0.17.0
yt-dlp==2023.10.13
pydantic==2.4.2
python-multipart==0.0.6
numpy==1.26.1 
datasets==3.5.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
groq==0.22.0
huggingface-hub==0.30.2
idna==3.10
jamo==0.4.1
Levenshtein==0.27.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
packaging==24.2
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==78.1.0
SQLAlchemy==2.0.40
tokenizers==0.21.1
transformers==4.51.3
urllib3==2.4.0
youtube-transcript-api==1.0.3